requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
dataclasses>=0.6
typing>=3.7.4.3
python-dotenv==1.0.0
typing-extensions==4.8.0
black==23.11.0
mypy==1.7.0
pytest-cov==4.1.0 